requirements.txt
pillow
wxPython~=4.2.2